updating to rev 195

+++ src/com/p3achb0t/api/interfaces/Client.kt
@@ -2,7 +2,7 @@ package com.p3achb0t.api.interfaces
 
 interface Client : GameShell {
     fun getVarbit(id: Int): Int
-    fun doAction(actionParam: Int, widgetID: Int, menuAction: Int, id: Int, menuOption: String, menuTarget: String, mouseX: Int, mouseY: Int, dummy: Byte=0)
+    fun doAction(actionParam: Int, widgetID: Int, menuAction: Int, id: Int, menuOption: String, menuTarget: String, mouseX: Int, mouseY: Int, dummy: Int=0)
     fun setScene_selectedX(x: Int)
     fun setScene_selectedY(y: Int)
     fun setViewportWalking()
