+++ src/com/p3achb0t/detours/Detours.java
@@ -15,7 +15,7 @@ public class Detours {
                                 String menuTarget,
                                 int mouseX,
                                 int mouseY,
-                                byte dummy,
+                                int dummy,
                                 InstanceManager scriptManager){
 
         ActionScript script = scriptManager.getActionScript();
